% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIMIC-package.r
\docType{package}
\name{BMIMIC-package}
\alias{BMIMIC-package}
\alias{BMIMIC}
\title{BMIMIC}
\description{
The BMIMIC package contains the data generation codes for
 paper.
}
\details{
The 'BMIMIC' package.
}
\references{
Sooyong Lee, Suhwa Han, Seung W. Choi (2023). BMIMIC.
}
\author{
Sooyong Lee
}
\keyword{package}
